add some more tests
